+++ src/components/alert-dialog.tsx
@@ -1,11 +1,18 @@
 'use client';
 
-import * as React from 'react';
 import { AlertDialog as AlertDialogPrimitive } from 'radix-ui';
+import * as React from 'react';
 
 import { cn } from '@/lib/utils';
 import { Button } from '@/src/components/button';
 
+/**
+ * This function renders an Alert Dialog component for the elementa-ui.
+ *
+ * @version 0.1.0
+ * @author Aayush Goyal
+ * @modifiedAt 2026-04-20
+ */
 function AlertDialog({
     ...props
 }: React.ComponentProps<typeof AlertDialogPrimitive.Root>) {